implemented with ring buffers
Ringbuffer was taken from somewhere. Made a change in the ringbuffer library to allow readers to be registered even though writers are already writing. Dunno why this was a feature of the original but if a writer already has begun writing, no more readers can read, obviously this is not going to wrok in a icecast/streaming setting.