spent hours chasing a bug realated to orphaned readaers
If you don't remove dead readers from the ring buffer, the writer throws an exception each time it wraps around the ring because that reader is not moving (dead) and it forces all the readers to catch up causing skips and hisses for the rest of the living readers. A reader can die if the consumer thread crashes and orphans the reader.